+++ CHANGES.md
@@ -5,6 +5,7 @@ This changelog tracks features and fixes of jskos PHP library.
 ## NEXT
 
 * Fix method Set->findURI to actually return an index
+* Add optional types parameter to jsonLDSerialize
 
 ## 0.3.2